fix ssl shutdown

+++ src/ssl/ssl.ts
@@ -10,50 +10,65 @@ export const httpsOptions = async () => {
     const keyPathOld = path.join(sslDirPath, 'privkey_old.pem');
     const certPathOld = path.join(sslDirPath, 'cert_old.pem');
 
-    if ((!fs.existsSync(keyPath) || !fs.existsSync(certPath)) && fs.existsSync(keyPathOld) && fs.existsSync(certPathOld)) {
-        if (fs.existsSync(certPath)) {
-            fs.rmSync(certPath);
-        }
-        if (fs.existsSync(keyPath)) {
-            fs.rmSync(keyPath);
-        }
-        fs.renameSync(keyPathOld, keyPath);
-        fs.renameSync(certPathOld, certPath);
-    }
-
-    const defaultCerts = {
-      key: fs.readFileSync(keyPath),
-      cert: fs.readFileSync(certPath),
-    };
-
-    fs.renameSync(keyPath, keyPathOld);
-    fs.renameSync(certPath, certPathOld);
-
-    const Downloader = require("nodejs-file-downloader");
-    const downloaderKey = new Downloader({
-        url: environment.security.keyUrl,
-        directory: sslDirPath, //This folder will be created, if it doesn't exist.   
-    });
-    const downloaderCert = new Downloader({
-      url: environment.security.certUrl,
-      directory: sslDirPath, //This folder will be created, if it doesn't exist.   
-    });
-    
     try {
+        // si une erreur est presente au demarrage rechercher les anciens certificats.
+        if ((!fs.existsSync(keyPath) || !fs.existsSync(certPath)) && fs.existsSync(keyPathOld) && fs.existsSync(certPathOld)) {
+            if (fs.existsSync(certPath)) {
+                fs.rmSync(certPath);
+            }
+            if (fs.existsSync(keyPath)) {
+                fs.rmSync(keyPath);
+            }
+            if (fs.existsSync(keyPathOld)) {
+                fs.renameSync(keyPathOld, keyPath);
+            }
+            if (fs.existsSync(certPathOld)) {
+                fs.renameSync(certPathOld, certPath);
+            }
+        }
+
+        fs.renameSync(keyPath, keyPathOld);
+        fs.renameSync(certPath, certPathOld);
+
+        const Downloader = require("nodejs-file-downloader");
+        const downloaderKey = new Downloader({
+            url: environment.security.keyUrl,
+            directory: sslDirPath, //This folder will be created, if it doesn't exist.   
+        });
+        const downloaderCert = new Downloader({
+            url: environment.security.certUrl,
+            directory: sslDirPath, //This folder will be created, if it doesn't exist.   
+        });
+
+        // Download of the new key/cert
         await downloaderKey.download();
         await downloaderCert.download();
 
-        fs.rmSync(keyPathOld);
-        fs.rmSync(certPathOld);
+        if (fs.existsSync(keyPathOld)) {
+            fs.rmSync(keyPathOld);
+        }
+        if (fs.existsSync(certPathOld)) {
+            fs.rmSync(certPathOld);
+        }
 
         return {
             key: fs.readFileSync(keyPath),
             cert: fs.readFileSync(certPath),
         };
     } catch (error) {
-        fs.renameSync(keyPathOld, keyPath);
-        fs.renameSync(certPathOld, certPath);
+        if (fs.existsSync(keyPathOld)) {
+            fs.renameSync(keyPathOld, keyPath);
+        }
+        if (fs.existsSync(certPathOld)) {
+            fs.renameSync(certPathOld, certPath);
+        }
         console.log("Download failed", error);
+
+        const defaultCerts = {
+            key: fs.readFileSync(keyPath),
+            cert: fs.readFileSync(certPath),
+        };
+        
         return defaultCerts;
     }
 };
